requirements.txt
alembic==1.8.0
Babel==2.10.1
elasticsearch==8.2.3
email-validator==1.2.1
Flask==2.1.2
Flask-Babel==2.0.0
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-Mail-SendGrid==0.3
Flask-Migrate==3.1.0
Flask-Moment==1.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
googletrans==4.0.0rc1
jinja2==3.0.3
PyJWT==2.4.0
pylint==2.14.2
python-dotenv==0.20.0
redis==4.3.4
requests==2.28.0
rq==1.11.0
SQLAlchemy==1.4.37
WTForms==3.0.1

pytest~=7.1.2
Werkzeug~=2.1.2
click~=8.1.3

# For heroku
psycopg2-binary==2.9.3
gunicorn==20.1.0